@@ -1,13 +1,14 @@
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_storyboard/flutter_storyboard.dart';
 import 'package:random_color/random_color.dart';
 
 void main() {
-  runApp(MyApp());
+  runApp(const MyApp());
 }
 
 class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+
   @override
   Widget build(BuildContext context) {
     return StoryBoard.material(
@@ -68,14 +69,15 @@ class MyApp extends StatelessWidget {
 }
 
 Widget _generateScreen({
-  Text title,
-  FloatingActionButton fab,
-  Color color,
+  Text? title,
+  FloatingActionButton? fab,
+  Color? color,
 }) {
   return Builder(
     builder: (context) {
-      final Map<String, dynamic> args =
-          ModalRoute.of(context).settings.arguments;
+      final route = ModalRoute.of(context);
+      final settings = route?.settings;
+      final args = settings?.arguments as Map<String, dynamic>?;
       return Scaffold(
         appBar: AppBar(title: title),
         backgroundColor: color,
@@ -87,7 +89,7 @@ Widget _generateScreen({
 }
 
 class HomeScreen extends StatelessWidget {
-  const HomeScreen({Key key}) : super(key: key);
+  const HomeScreen({super.key});
 
   @override
   Widget build(BuildContext context) {
@@ -112,7 +114,7 @@ class HomeScreen extends StatelessWidget {
 }
 
 class CounterScreen extends StatefulWidget {
-  const CounterScreen({Key key}) : super(key: key);
+  const CounterScreen({super.key});
 
   @override
   _CounterScreenState createState() => _CounterScreenState();
