@@ -3,10 +3,12 @@ import 'package:flutter_storyboard/flutter_storyboard.dart';
 import 'package:random_color/random_color.dart';
 
 void main() {
-  runApp(MyApp());
+  runApp(const MyApp());
 }
 
 class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+
   @override
   Widget build(BuildContext context) {
     return StoryBoard(
@@ -18,23 +20,24 @@ class MyApp extends StatelessWidget {
           margin: const EdgeInsets.all(4.0),
           color: RandomColor(title.hashCode).randomColor(),
           child: Stack(
-            overflow: Overflow.visible,
+            clipBehavior: Clip.none,
             children: [
               child,
-              Positioned(
-                left: -50,
-                top: 350,
-                child: RotatedBox(
-                  quarterTurns: 3,
-                  child: Text(
-                    title,
-                    style: TextStyle(
-                      color: Colors.black,
-                      fontSize: 30,
+              if (title != null)
+                Positioned(
+                  left: -50,
+                  top: 350,
+                  child: RotatedBox(
+                    quarterTurns: 3,
+                    child: Text(
+                      title,
+                      style: TextStyle(
+                        color: Colors.black,
+                        fontSize: 30,
+                      ),
                     ),
                   ),
                 ),
-              ),
             ],
           ),
         );
@@ -83,14 +86,15 @@ class MyApp extends StatelessWidget {
 }
 
 Widget _generateScreen({
-  Text title,
-  FloatingActionButton fab,
-  Color color,
+  Text? title,
+  FloatingActionButton? fab,
+  Color? color,
 }) {
   return Builder(
     builder: (context) {
-      final Map<String, dynamic> args =
-          ModalRoute.of(context).settings.arguments;
+      final route = ModalRoute.of(context);
+      final settings = route?.settings;
+      final args = settings?.arguments as Map<String, dynamic>?;
       return Scaffold(
         appBar: AppBar(title: title),
         backgroundColor: color,
