@@ -3,17 +3,19 @@ import 'package:flutter_storyboard/flutter_storyboard.dart';
 import 'package:random_color/random_color.dart';
 
 void main() {
-  runApp(MyApp());
+  runApp(const MyApp());
 }
 
 class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+
   @override
   Widget build(BuildContext context) {
     return StoryBoard(
       showAppBar: true,
-      children: [
+      sizedChildren: [
         for (var i = 4; i < 20; i++)
-          SizedBox.fromSize(
+          CustomScreen(
             size: Size(i * 50.0, i * 100.0),
             child: _generateScreen(
               title: Text('Screen$i'),
@@ -21,9 +23,9 @@ class MyApp extends StatelessWidget {
             ),
           ),
       ],
-      sizedChildren: [
+      children: [
         for (var i = 4; i < 20; i++)
-          CustomScreen(
+          SizedBox.fromSize(
             size: Size(i * 50.0, i * 100.0),
             child: _generateScreen(
               title: Text('Screen$i'),
@@ -36,14 +38,15 @@ class MyApp extends StatelessWidget {
 }
 
 Widget _generateScreen({
-  Text title,
-  FloatingActionButton fab,
-  Color color,
+  Text? title,
+  FloatingActionButton? fab,
+  Color? color,
 }) {
   return Builder(
     builder: (context) {
-      final Map<String, dynamic> args =
-          ModalRoute.of(context).settings.arguments;
+      final route = ModalRoute.of(context);
+      final settings = route?.settings;
+      final args = settings?.arguments as Map<String, dynamic>?;
       return Scaffold(
         appBar: AppBar(title: title),
         backgroundColor: color,
